refactor: move PanoramaTileProcessing internals to PanoramaLayer
To continue the removal of some Processing files in Process, all
contents of PanoramaTileProcessing have been moved to PanoramaLayer.
